pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Yuuki Fujihara <mail@fujiy.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
protobuf = "^3.17.0"
grpcio = "^1.38.0"
grpcio-tools = "^1.38.0"
Pillow = "^8.2.0"
open3d = "^0.12.0"
numpy = "^1.20.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
